requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
email-validator==2.1.0
resend==0.7.0
sqlalchemy==2.0.23
werkzeug==3.0.1
python-dotenv==1.0.1